ver 2.0

+++ src/main/java/com/mycom/board/BoardController.java
@@ -6,6 +6,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.List;
 
 
 @Controller
@@ -14,14 +15,11 @@ public class BoardController {
     BoardService boardService;
 
     @RequestMapping(value = "/getListBoard.do")
-    public String getListBoard(Model model1, Model model2, HttpServletRequest request, BoardVO vo) {
+    public String getListBoard(Model model, HttpServletRequest request, BoardVO vo) {
         System.out.println("===>Controller로 getListBoard() 접속");
 
         int currentPageNo = 0;
         int currentRecord = 0;
-
-        System.out.println("ch1 : "+vo.getCh1());
-        System.out.println("ch2 : "+vo.getCh2());
 
         if(request.getParameter("currentPageNo")!=null) {
             currentPageNo = (Integer.parseInt(request.getParameter("currentPageNo")));
@@ -32,17 +30,26 @@ public class BoardController {
             }
         }
 
+        if(request.getParameter("ch1")!=null){
+            vo.setCh1(request.getParameter("ch1"));
+        }
+        if(request.getParameter("ch2")!=null){
+            vo.setCh2(request.getParameter("ch2"));
+        }
+
+        System.out.println("ch1 : "+vo.getCh1());
+        System.out.println("ch2 : "+vo.getCh2());
+
         boardService.getListBoard(vo);
         BoardPaging boardPaging = new BoardPaging(boardService.getTotalCount(), currentPageNo);
 
         vo.setMinLimit(currentRecord);
         vo.setMaxLimit(boardPaging.getPageSize());
 
-        System.out.println("min"+vo.getMinLimit());
-
-
-        model1.addAttribute("boardList", boardService.getListBoard(vo));
-        model2.addAttribute("boardPage", boardPaging);
+        model.addAttribute("boardList", boardService.getListBoard(vo));
+        model.addAttribute("boardPage", boardPaging);
+        model.addAttribute("ch1", vo.getCh1());
+        model.addAttribute("ch2", vo.getCh2());
         return "/board/listBoard.jsp";
     }
 
@@ -54,9 +61,15 @@ public class BoardController {
         return "/board/contentBoard.jsp";
     }
 
-    @RequestMapping(value = "/getUpdateContent.do")
-    public String getUpdateContent(BoardVO vo, Model model){
+    @RequestMapping(value = "/getUpdateBoardContent.do")
+    public String getUpdateContent(BoardVO vo, Model model, HttpServletRequest request){
         System.out.println("===>Controller로 getUpdateContent() 접속");
+        BoardVO m = boardService.getBoard(vo);
+        String passwd = request.getParameter("passwd");
+        if(!passwd.equals(m.getPasswd())){
+            model.addAttribute("num", vo.getNum());
+            return "/board/errorPasswd.jsp?num="+vo.getNum()+"";
+        }
         model.addAttribute("board",boardService.getBoard(vo));
         return "/board/updateBoard.jsp";
     }
@@ -76,8 +89,15 @@ public class BoardController {
     }
 
     @RequestMapping(value = "/deleteBoard.do")
-    public String deleteBoard(BoardVO vo){
+    public String deleteBoard(BoardVO vo, HttpServletRequest request, Model model){
         System.out.println("===>Controller로 deleteBoard() 접속");
+        BoardVO m = boardService.getBoard(vo);
+        String passwd = request.getParameter("passwd");
+        if(!passwd.equals(m.getPasswd())){
+            System.out.println("getupdate 접똑!");
+            model.addAttribute("num", vo.getNum());
+            return "/board/errorPasswd.jsp?num="+vo.getNum()+"";
+        }
         boardService.deleteBoard(vo);
         return "redirect:getListBoard.do";
     }
